feat: Add local storage for editor persistence
Adds editor state serialization and local storage for code and settings (volume).
The stored code can be reset by deleting all of it and refreshing the page. Later
this will need a UI to do this more explicitly.

Additionally, the app is refactored a bit to improve readability (moving demo
code into a separate module, dealing with volume consistently in the range
of 0-1, using an interface for cursor location).