feat: Add gain property for samples
Updates the parser to support the unary minus operator, thereby implementing
negative numbers (both for literals, and more complex expressions). Adds `gain`
as an argument of the `sample()` function, along with the necessary audio engine
changes to let it take effect. Finally, updates the demo to include a -3 dB gain
on one of the instruments.