feat: Play only once and sync state from AudioEngine
This patch introduces an Observable interface, which is then used to make the
AudioEngine the source of truth for "playing" and "volume" data, instead of
the UI layer holding that data. Most importantly, this allows us to detect when
the playback is done in the engine and easily set "playing" back to false. The
previous behavior was to loop indefinitely.

Also included in this patch are the removal of the "Update" button, which has
become obsolete (arguably) now that the track no longer loops, and a refactor
of pitch-to-MIDI mapping into its own module.