feat: Support multiple routings per instrument (#66)
This patch lifts the requirement that per section, only one pattern may be
routed to an instrument. As an example, the following is now possible:

```
section main for 4 bars {
  kick << [x--- ----]
  kick << [---- x---]
}
```

Both routings would create their own independent sequences, in effect playing
the kickdrum on every beat.

How overlapping step information is handled would depend on the instrument.