feat: Add mixer routing support to audio engine (#67)
See the updated demo code in src/ui/demo.ts for the new possibilities.

This patch adds a "gain" property (in decibels) to bus objects. Further, the mixer
now supports routing statements, where the destination is a bus and the source
is either an instrument or another bus.

This creates a graph of instrument and bus nodes. That graph is directly used
in the audio engine to set up a graph of Tone.js audio nodes. Any instrument
or bus not connected to anything is automatically sent to the main output.
Gain is applied on bus nodes according to the "gain" property.