feat: Add tabbed layout and settings page
This patch adds a TabLayout component using @headlessui/react for tabbed
navigation. Right now, there are (always) an "Editor" tab (previously the only
available view) and a "Settings" tab. The settings page adds the ability to reset
the project and load the demo in a non-hacky way.

Some layout changes are done throughout the app to save space.