feat: Add Mixer page with routing visualization
Adds a page that shows the mixer nodes (instruments, buses, and the main
output) as well as mixer edges (routings) as a graph.

This is intended to help new users to learn the routing mechanics, and also to
help debug issues with complex graphs and to check for e.g. missing routings.