feat: Add timeline page (#90)
Implements a basic timeline component and adds it to a new "Timeline" page.
The component includes a ruler, the track sections, and a progress indicator
while the track is playing. The mousewheel can be used to zoom.

In the future, this would be better positioned as a panel docked below the
editor (collapsible of course), but the UI layout system needs some more work
to enable that.