feat: Redesign settings UI
This patch is both a refactor and a redesign of the settings page. Feature-wise,
an output gain slider is added to the settings, as well as a theme selection
with options for dark (default), light, and system theme. However, dark and
system are still disabled and will be implemented in a future patch.