feat: Implement layout persistence
The layout structure is now fully serializable and no longer includes functions,
instead referring to elements via a string type (and optional props, for future
extensibility). Rendering of the tab title and content is then handled by a
registry-like module.

Initially, the layout is taken from local storage (or initialized from a constant).
A set of React reducer/dispatch functions handles updates. These include
changes to the selected tab, as well as resizing split panes, and will be extended
to reordering of tabs, splitting and merging of panes, etc., once these features
are implemented.

Due to the storage data getting more extensive, saving to storage is now also
debounced.