feat: Add keyboard shortcuts for commands (#112)
The command palette displays the keyboard shortcut next to search
results. It also handles dispatching of commands when the shortcut is
pressed.

Ctrl+Shift+Space is added for toggling playback. Ctrl+Space would be
less awkward, but if/when code completion is added, then that should get
Ctrl+Space instead.