feat: Implement drag-and-drop for tabs (#122)
Tabs can now be dragged to reorder, as well as to move them between
different panes (tab lists). While doing so, a preview of the changed
layout is rendered.

It was required to make layout dispatch an optional prop everywhere for
compatibility between the preview and any external changes, e.g. due to
tab focusing by HeadlessUI.