feat: Preview tab DnD abstractly instead of recomputing layout (#125)
We no longer compute a full preview layout during drag-and-drop of tabs.
Firstly, this is somewhat expensive, and secondly, it gets complicated
once splitting and merging of panes is introduced. Instead, show a line
marker where the tab will end up. This is also what VS Code does, and it
works nicely.