feat: Implement splitting panes by drag-and-drop
Dragging tabs into the north/south/east/west or center region of a pane
will now highlight that area, and when dropped, will do the following:

* For the center region, will add the dragged tab into the pane, next
    to the currently active tab in that pane.
* For the other regions, will replace the pane with a split node that
    contains the pane as well as another pane for only the dragged tab.
    In effect, this splits the pane in half, allocating the other half
    to the dragged tab.

A normalization function ensures that unneeded (empty) panes are removed
and the layout is as flat as possible.

The 1px border at the top of tab lists is removed, and the separation
is instead handled by the SplitNodeView's resize handle, to ensure there
is also a border between vertically split nodes.