fix: Update react-resizable-panels and prevent change cycles (#136)
Updates the react-resizable-panels dependency to v4 and applies the
necessary syntax changes. Since we need a controlled component, but the
logic for this no longer works with v4, a full rework is needed.

Some peculiar behavior can be observed with the new version. When the
layout is changed significantly by us, the library sometimes emits
multiple onLayoutChange events, some of which with an incomplete set of
panels (e.g. only the first) and with sizes that do not correspond to
what we set. This is either a bug in the library or a misunderstanding
on our side (still unclear, despite thorough investigation).

To mitigate this, we keep track of the initial source of the change. For
changes that we initiated, we ignore onLayoutChange, and if the sizes
differ, re-assert the expected layout.