feat: Support range selection in timeline (#138)
When the user performs a plain left click on the timeline ruler, this
still just sets the start position. However, when the user holds down
the mouse button and drags the mouse left or right, this will now create
a range selection with start and end markers, and any playback will be
limited to that selection, meaning it will begin at the first marker and
end at the second marker.