feat: Add bus effects, with first "delay" effect (#149)
This patch adds support for adding audio effects on buses. The effects
are processed before the bus gain is applied. It is possible to specify
multiple effects. The syntax is simply `effect <expression>` where
`<expression>` can be anything that is of type `effect`, such as the
result of a constructor function. One such function is added with this
patch: `delay(time: number<steps>, feedback: number)`, which creates a
feedback delay effect.