refactor: Let pattern interface support infinite patterns (#158)
This does not impact the observable behavior of the language, but it
lays the foundation for upcoming changes to how patterns work. Instead
of patterns always being a finite series of steps, they can now be
either finite or infinite, and must be evaluated to render a specific
number of steps. This makes it possible to implement concatenation and
multiplication without copying the pattern data.

The goal is (in a future patch) to remove the implicit looping behavior
of track sections. Instead, patterns will be played as-is, and must be
explicitly looped (via something like `loop([x---]`), which would make
them infinite and fill the section.