feat!: Require explicit pattern looping to fill sections
Before this patch, sections would implicitly repeat any pattern within
to fill the length of the section. While this is helpful for building
drum loops, it can be unexpected for new users, and makes it difficult
to trigger one-shot sounds.

Therefore, sections will now add rest steps after patterns that are too
short, instead of looping them. To regain the looping behavior, a new
function `loop(<pattern>)` is added that returns an infinitely repeating
version of its argument.